App Update
Setup is now 2 stages with ability to skip it all.
Profile is cleaned up slightly with automatic refresh of changes made to
items.
Watchlist is now a private class within main that gets a name for the
user rather than asking them for one initially the name can be changed
in profile.